requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.54.0
chromadb>=0.5.20
pymupdf>=1.24.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
tiktoken>=0.7.0
slowapi>=0.1.9
